# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Almáskosár API",
        "description": "Az Almáskosár Webshop API végpontjai",
        "version": "2.0.0"
    },
    "paths": {
        "/api/carts": {
            "get": {
                "tags": [
                    "Cart"
                ],
                "summary": "Listázza az összes kosarat",
                "description": "Listázza az összes kosarat",
                "operationId": "getCartsList",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CartsSchema"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Products not found"
                    }
                }
            },
            "post": {
                "tags": [
                    "Cart"
                ],
                "summary": "Új kosár létrehozása",
                "description": "Új kosár létrehozása",
                "operationId": "2176b5481799873176eccc4def6e8ca7",
                "requestBody": {
                    "description": "Új kosár adatai",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CartsSchema"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "A kosár sikeresen létrejött",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "Az üzenet",
                                            "type": "string",
                                            "example": "Sikeresen hozzáadva!"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/carts/{id}/edit": {
            "get": {
                "tags": [
                    "Cart"
                ],
                "summary": "Kosár keresése",
                "description": "Visszaadja a keresett kosár adatait",
                "operationId": "a60b90ddd7b4f016fc96e35fb3e7c4ab",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Keresett kosár azonosítója",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Visszaadja a keresett kosarat",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CartsSchema"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Product not found"
                    }
                }
            }
        },
        "/api/carts/{id}": {
            "put": {
                "tags": [
                    "Cart"
                ],
                "summary": "Kosár szerkesztése",
                "description": "Frissíti a keresett kosár adatait",
                "operationId": "updateCart",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Frissíteni kívánt kosár azonosítója",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CartsSchema"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Product updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CartsSchema"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid input"
                    },
                    "404": {
                        "description": "Product not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Cart"
                ],
                "summary": "Kosár törlése",
                "description": "Törli a keresett kosarat",
                "operationId": "1ec691dc69823fc0a41f12762eea2292",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Törölni kívánt termék azonosítója",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Cart deleted successfully"
                    },
                    "404": {
                        "description": "Cart not found"
                    },
                    "500": {
                        "description": "Server Error"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "CartsSchema": {
                "properties": {
                    "id": {
                        "description": "Kosár azonosítója",
                        "type": "integer",
                        "example": "1",
                        "nullable": false
                    },
                    "user_id": {
                        "description": "Felhasználó azonosítója",
                        "type": "integer",
                        "example": "5",
                        "nullable": false
                    },
                    "category_id": {
                        "description": "Termék kategóriájának azonosítója",
                        "type": "string",
                        "example": "2",
                        "nullable": false
                    },
                    "product_id": {
                        "description": "Termék azonosítója",
                        "type": "string",
                        "example": "14",
                        "nullable": false
                    },
                    "product_quantity": {
                        "description": "Termék mennyisége",
                        "type": "string",
                        "example": "2",
                        "nullable": false
                    }
                },
                "type": "object"
            },
...
        }
    }
}